harmonize token & cookie expiration time
Previously, we would store the creation time of the token,
and apply the "max age" during decoding. This is in contrast
to how cookies work. A cookie's max-age is fixed at the time
it's created.

After this patch, the token's expiration time will be fixed
during encoding, so it matches how the cookie works.

This only makes a difference if the server operator changes
the configured max age. We want that configuration change to
apply only to newly-created tokens, just like it only
applies to newly-created cookies.